chore: restructure WidgetsEditor
