+++ app/client/src/pages/Editor/IDE/RightPane/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PropertyPaneWrapper from "pages/Editor/WidgetsEditor/PropertyPaneWrapper";
+import PropertyPaneWrapper from "pages/Editor/WidgetsEditor/components/PropertyPaneWrapper";
 import {
   ADD_PATH,
   BUILDER_CUSTOM_PATH,
