+++ app/client/src/pages/Editor/WidgetsEditor/components/PropertyPaneWrapper.tsx
@@ -3,7 +3,7 @@ import PropertyPaneSidebar from "components/editorComponents/PropertyPaneSidebar
 import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { getPropertyPaneWidth } from "selectors/propertyPaneSelectors";
-import { CreateNewModal } from "../IDE/RightPane/components/CreateNewModal";
+import { CreateNewModal } from "../../IDE/RightPane/components/CreateNewModal";
 
 /**
  * OldName: PropertyPaneContainer
